+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Hypothesis `database` setting from `@settings` decorator or loaded profiles being ignored with `@schema.parametrize()`.
 - `AttributeError` when using Schemathesis with HypoFuzz due to uninitialized `_override` attribute in `ProjectsConfig`.
 
+### :wrench: Changed
+
+- Bump `pytest-subtest` from <0.15.0 to <0.16.0.
+
 ## [4.7.1](https://github.com/schemathesis/schemathesis/compare/v4.7.0...v4.7.1) - 2025-12-07
 
 ### :bug: Fixed